+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 0.13.1.9000
+# targets 0.13.2
 
 ## Bug fixes